+++ src/api/annonars.py
@@ -5,7 +5,7 @@ from pydantic import ValidationError
 
 from src.core.config import settings
 from src.defs.annonars import AnnonarsRangeResponse
-from src.seqvar import SeqVar
+from src.defs.seqvar import SeqVar
 
 #: Annonars API base URL
 ANNONARS_API_BASE_URL = f"{settings.API_REEV_URL}/annonars"
