+++ src/api/dotty.py
@@ -5,7 +5,7 @@ from pydantic import ValidationError
 
 from src.core.config import settings
 from src.defs.dotty import DottySpdiResponse
-from src.genome_builds import GenomeRelease
+from src.defs.genome_builds import GenomeRelease
 
 #: Dotty API base URL
 DOTTI_API_BASE_URL = f"{settings.API_REEV_URL}/dotty"
