+++ src/defs/seqvar.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 from src.api.dotty import DottyClient
 from src.core.exceptions import InvalidPos, ParseError
-from src.genome_builds import (
+from src.defs.genome_builds import (
     CHROM_LENGTHS_37,
     CHROM_LENGTHS_38,
     REFSEQ_CHROM_37,
