+++ tests/api/test_annonars.py
@@ -2,8 +2,8 @@ import responses
 
 from src.api.annonars import AnnonarsClient
 from src.defs.annonars import AnnonarsRangeResponse
-from src.genome_builds import GenomeRelease
-from src.seqvar import SeqVar
+from src.defs.genome_builds import GenomeRelease
+from src.defs.seqvar import SeqVar
 from tests.utils import get_json_object
 
 #: Example sequence variant
